Extract updater prerequisite validation
