+++ apps/server/vibesensor/use_cases/updates/__init__.py
@@ -8,8 +8,10 @@ Module topology
 - **Facade**: ``manager.py`` — public ``UpdateManager`` API for routes and
   runtime lifecycle.  Update workflow orchestration is inlined in
   ``_run_update_inner()``.
-- **Validation & services**: ``workflow.py`` — prerequisite validation and
-  service control.
+- **Validation**: ``validation.py`` — runtime prerequisite checks for tools,
+  privileges, rollback storage, and disk space before update orchestration.
+- **Services**: ``manager.py`` — backend restart scheduling and runtime update
+  lifecycle orchestration.
 - **Operations**: ``installer.py`` (install/rollback), ``wifi.py`` (Wi-Fi
   connect/restore, diagnostics, network constants), ``releases.py``
   (GitHub release discovery), ``runner.py`` (process execution and
